clk: tegra: Support runtime PM, power domain and OPP
The Clock-and-Reset controller resides in a "core" power domain on
NVIDIA Tegra SoCs. In order to support voltage scaling of the core power
rail, we need to hook up some clocks (which can't operate properly on a
lower voltages above certain clock rates) to the core power domain in
order to ensure that a proper core voltage is maintained.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>