mm: Add kmem_cache_last_alloc() arguments for stack trace
In some cases, the allocator return address is in a common function,
so that more information is desired.  This commit therefore adds stackp
and nstackp parameters to kmem_cache_last_alloc() in order to allow
CONFIG_STACKTRACE=y slub stack traces to be obtained.  This will be used
in a later commit that will provide a kmem_last_alloc_stack() function.

Cc: Christoph Lameter <cl@linux.com>
Cc: Pekka Enberg <penberg@kernel.org>
Cc: David Rientjes <rientjes@google.com>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Andrew Morton <akpm@linux-foundation.org>
Cc: <linux-mm@kvack.org>
Reported-by: Andrii Nakryiko <andrii@kernel.org>
[ paulmck: Move slab definition per Stephen Rothwell and kbuild test robot. ]
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>