btrfs: abort the transaction if we fail to inc ref in btrfs_copy_root
I hit a pretty ugly corruption when doing some error injection, it turns
out this is because we do not abort the transaction if we fail to adjust
the reference counts in btrfs_copy_root().  This makes us miss updating
the references for the new root that has already been allocated and
inserted, which is the problem.  Fix this by aborting the transaction if
we fail to do the reference modification.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>
Signed-off-by: David Sterba <dsterba@suse.com>