WIP: drm/grate: Add Host1x and DRM drivers with experimental changes
The new drivers and UAPI provide all features that are necessary for
implementing of a proper userspace driver. The new driver takes a different
approach in regards to working with sync points by making each submitted job
to take an individual sync point, this solves the problem with the sync point
recovery that old driver suffered from. The new Host1x driver and v2 UAPI are
optimized for performance and minimal resources consumption. The new v2 UAPI
exposes to userspace the HW cmdstream-related features that are necessary
for a proper HW fencing. The staging v1 UAPI is kept functional and all of
current userspace will continue to work with the new driver, note that some
of the never-really-used features of v1 UAPI have been removed (like sync
point increment IOCTL) and will return EPERM.

The goal is to merge new features into the main driver once they are ready.

Key moments of v2 UAPI:
	- Raw sync points are not exposed to userspace.

	- Job descriptors are embedded into the commands stream. Kernel
	  driver parses the stream and patches the descriptors in-place.

	- Commands buffer is copied from userspace via usrptr (no BO
	  allocation and uncached-reading overhead).

	- Channel is not restricted to a single client, thus 3d channel can
	  take a multi-client job that uses 2d / 3d.

	- Supports explicit jobs fencing. DRM sync object and DRM scheduler
	  are utilized to provide the fencing support.

	- Allows to use host1x gather opcode, which is restricted to
	  data-upload usecase only. This allows userspace to pre-allocate
	  a gather buffer, put shaders code / data there, and use two-word
	  "gather" opcode to upload data from the buffer without pushing
	  that data into the commands stream on each submission.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>