dt-bindings: clock: tegra-car: Document new tegra-clocks node
Document tegra-clocks sub-node which describes Tegra SoC clocks that
require a higher voltage of the core power domain in order to operate
properly on a higher clock rates.  Each node contains a phandle to OPP
table and to power domain, which are used for scaling voltage based on
the clock rate.

Some of the root Tegra SoC PLLs don't have any specific device which
controls the clock state and the Clock controller is in charge of these
clocks.  Hence there is a need to specify OPP table for these clocks in
order perform DVFS on them.

Some clocks have a specific peripheral device that is associated with the
clock, but these clocks don't require very high voltages and it's much
more convenient that Clock controller could manage the voltage state for
them, rather than to put this burden on each individual device.

This binding describes the clocks as sub-devices of the Clock controller.

Tested-by: Peter Geis <pgwipeout@gmail.com>
Tested-by: Nicolas Chauvet <kwizart@gmail.com>
Tested-by: Matt Merhar <mattmerhar@protonmail.com>
Signed-off-by: Dmitry Osipenko <digetx@gmail.com>