ARM: tegra: fix emmc on specific bootloaders
Ouya fails to detect the emmc module when booting from specific bootloaders.
Fastboot and hard-kexec bootloaders have observed failures while u-boot does not have this issue.
It was discovered that if the sdmmc4 alternate configuration clock pin is input disabled the issue manifests.
Ouya uses sdmmc4 in the primary pin configuration so it is unknown why this occurs.
It is likely related to the other emmc limitations we have on Ouya.

For now, fix it by enabling input on cam_mclk_pcc0.

Fixes: d7195ac5c9c5 ("ARM: tegra: Add device-tree for Ouya")
Signed-off-by: Peter Geis <pgwipeout@gmail.com>
Reported-by: Matt Merhar <mattmerhar@protonmail.com>
Tested-by: Matt Merhar <mattmerhar@protonmail.com>