ARM: 9052/1: decompressor: cover BSS in cache clean and reorder with MMU disable on v7
Commit 401b368caaec ("ARM: decompressor: switch to by-VA cache maintenance
for v7 cores") replaced the by-set/way cache maintenance in the decompressor
with by-VA cache maintenance, which is more appropriate for the task at
hand, especially under virtualization on hosts with non-architected system
caches that are not affected by by-set/way maintenance at all.

On such systems, that commit inadvertently removed the cache clean and
invalidate of all of the guest's memory that is performed by KVM on behalf
of the guest after its MMU is disabled (but only if any by-set/way cache
maintenance instructions were issued first). This resulted in various
erroneous behaviors observed by Russell, all involving the mini-stack
used by the core kernel's v7 boot code, and which resides in BSS. It
seems intractable to figure out exactly what goes wrong in each of these
cases, but some small experiments did suggest that the lack of a cache
clean and invalidate *after* disabling the MMU and caches is what
triggers the errors, presumably because cachelines are being allocated
or reallocated while the first cache clean and invalidate is in progress.

To ensure that no cache lines cover any of the data that is accessed by
the booting kernel with the MMU off, include the uncompressed kernel's
BSS region in the cache clean operation.

Also, to ensure that no cachelines are allocated while the cache is being
cleaned, perform the cache clean operation *after* disabling the MMU and
caches when running on v7 or later, by making a tail call to the clean
routine from the cache_off routine. This requires passing the VA range
to cache_off(), which means some care needs to be taken to preserve
R0 and R1 across the call to cache_off().

Since this makes the first cache clean redundant, call it with the
range reduced to zero. This only affects v7, as all other versions
ignore R0/R1 entirely.

Link: https://lore.kernel.org/linux-arm-kernel/20210122152012.30075-1-ardb@kernel.org

Fixes: 401b368caaec ("ARM: decompressor: switch to by-VA cache maintenance for v7 cores")
Reported-by: Russell King <rmk+kernel@armlinux.org.uk>
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Signed-off-by: Russell King <rmk+kernel@armlinux.org.uk>