mm/compaction: fix edge case of fast_find_migrateblock()
In the fast_find_migrateblock(), It iterate freelist to find proper
pageblock.  But there are two edge cases.  First, if the page we found is
equal to cc->migrate_pfn, it is considered that we didn't found suitable
pageblock.  Second, if the loop was terminated because order is less than
PAGE_ALLOC_COSTLY_ORDER, it could be considered that we found suitable
one.

Link: https://lkml.kernel.org/r/20210123154320.24278-1-vvghjk1234@gmail.com
Fixes: 70b44595eafe9 ("mm, compaction: use free lists to quickly locate a migration source")
Signed-off-by: Wonhyuk Yang <vvghjk1234@gmail.com>
Cc: Mel Gorman <mgorman@techsingularity.net>
Cc: Vlastimil Babka <Vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>