powerpc: Rename kexec elfcorehdr_addr to elf_load_addr
The architecture specific field, elfcorehdr_addr in struct kimage_arch,
that holds the address of the buffer in memory for ELF core header for
powerpc has a different name than the one used for x86_64.  This makes
it hard to have a common code for setting up the device tree for
kexec system call.

Rename elfcorehdr_addr to elf_load_addr to align with x86_64 name so
common code can use it.

Signed-off-by: Rob Herring <robh@kernel.org>
Reviewed-by: Lakshmi Ramasubramanian <nramas@linux.microsoft.com>
Reviewed-by: Thiago Jung Bauermann <bauerman@linux.ibm.com>
Link: https://lore.kernel.org/r/20210213161049.6190-2-nramas@linux.microsoft.com
Signed-off-by: Rob Herring <robh@kernel.org>