drm/tegra: dc: Fix setting PCLK to 0Hz
RGB output doesn't permit to change parent clock rate of the display and
PCLK rate is set to 0Hz in this case. The tegra_dc_commit_state() shall
not set the display clock to 0Hz since this change propagates to the
parent clock. The DISP clock is defined as a NODIV clock by the tegra-clk
driver and all NODIV clocks use the CLK_SET_RATE_PARENT flag.

This bug stayed unnoticed because by default PLLP is used as the parent
clock of the display controller and PLLP silently skips the erroneous 0Hz
rate changes because it has other active child clocks that don't allow
rate changes. The PLLP isn't acceptable for some devices that we want to
upstream due to a display panel clock rate tolerance requirement which
can't be met by using PLLP and then the bug pops up in this case since
parent clock is set to 0Hz, killing the display output.

Instead of applying rate separately to the DC clock and its parent, apply
rate only to the DC clk and let CCF to propagate the change to the parent.
Now the 0Hz won't be set, the problem is fixed.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>