ASoC: tegra20: i2s: Add reset control
The I2S reset may be asserted at a boot time. Tegra20 I2S driver doesn't
manage the reset control and currently it happens to work because reset
is implicitly deasserted by the tegra-clk driver when I2S clock is
enabled. The I2S permanently stays in a reset once tegra-clk is fixed to
not touch the resets, which it shouldn't be doing. Add reset control to
the Tegra20 I2S driver.

Note that I2S reset was always specified in Tegra20 device-tree, hence
DTB ABI changes aren't required.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>