s390/smp: remove redundant pcpu::lowcore member
A pointer to CPU lowcore is stored in a struct cpu::lowcore
member and duplicated in lowcore_ptr[] global array. The
value of a lowcore is expected to be the same in both
locations and it is, except for the boot CPU.

The problem is after a CPU off/on hotplug cycle the value
in lowcore_ptr[0] (mistakenly) gets set to NULL, while the
pcpu::lowcore does not. That is inconsistent, but luckily
helps to avoid a NULL-pointer assignment, since pcpu::lowcore
is used in code paths where it could hit otherwise.

The boot CPU lowcore needs a special care already (it is
allocated with DAT off and re-used across CPU hotplug cycles).
Let's save some bits and make things more clear by simply
removing the superfluous pcpu::lowcore.

Signed-off-by: Alexander Gordeev <agordeev@linux.ibm.com>
Signed-off-by: Heiko Carstens <hca@linux.ibm.com>